tcr payoff implementation
This now has fully functional methods for hypothetical getting payoffs for all players involved, one payoff for each action - based on the actions of the other players.

We'll be able to use this to determine the best strategy for each player in a given scenario, which we can use to determine whether a given scenario is in a state of equilibrium.